#11 [MOVE] 컴포넌트 패키지로 분리
